requirements.txt
discord.py==2.3.2
aiogram==3.13.1
aiohttp==3.10.5